+++ easyrepo/__init__.py
@@ -1,4 +1,2 @@
 from easyrepo.interface.crud import CRUDRepository
 from easyrepo.interface.paging import PagingRepository
-from easyrepo.repository.mongo import MongoRepository
-from easyrepo.repository.memory import MemoryRepository